consolidate direct passive trace clearance
